replace the lookup in the plugin itself
